requirements.txt
openai
google-generativeai
Pillow
google-genai